Remove new test file; move two currently-failing tests to existing modules
